refactor: relocate main layout components
