+++ frontend/src/app/app.component.ts
@@ -1,7 +1,7 @@
 // frontend/src/app/app.component.ts
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MainLayoutComponent } from './layouts/main-layout/main-layout.component';
+import { MainLayoutComponent } from './layout/main-layout/main-layout.component';
 import { ModalComponent } from './components/ui/modal/modal.component';
 import { NotificationComponent } from './components/ui/notification/notification.component';
