+++ frontend/src/app/layout/components/main-layout/citation-footer/citation-footer.component.html
