+++ frontend/src/app/layout/components/main-layout/navigation/navigation.component.html
@@ -1,4 +1,4 @@
-<!-- frontend/src/app/layouts/main-layout/components/navigation/navigation.component.html -->
+<!-- frontend/src/app/layout/components/main-layout/navigation/navigation.component.html -->
 <header class="app-header">
   <div class="header-container">
     <div class="logo-container">
