+++ frontend/src/app/layout/components/main-layout/navigation/navigation.component.scss
@@ -1,4 +1,4 @@
-// frontend/src/app/layouts/main-layout/components/navigation/navigation.component.scss
+// frontend/src/app/layout/components/main-layout/navigation/navigation.component.scss
 
 // Glassmorphism Header
 .app-header {
