+++ frontend/src/app/layout/main-layout/main-layout.component.html
@@ -1,4 +1,4 @@
-<!-- frontend/src/app/layouts/main-layout/main-layout.component.html -->
+<!-- frontend/src/app/layout/main-layout/main-layout.component.html -->
 <div class="app-container">
   <app-navigation></app-navigation>
   <main class="app-content">
