+++ frontend/src/app/layout/main-layout/main-layout.component.scss
@@ -1,4 +1,4 @@
-// frontend/src/app/layouts/main-layout/main-layout.component.scss
+// frontend/src/app/layout/main-layout/main-layout.component.scss
 
 // App Container
 .app-container {
