+++ frontend/src/app/layout/main-layout/main-layout.component.ts
@@ -1,9 +1,9 @@
-// frontend/src/app/layouts/main-layout/main-layout.component.ts
+// frontend/src/app/layout/main-layout/main-layout.component.ts
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
-import { NavigationComponent } from './components/navigation/navigation.component';
-import { CitationFooterComponent } from './components/citation-footer/citation-footer.component';
+import { NavigationComponent } from '../components/main-layout/navigation/navigation.component';
+import { CitationFooterComponent } from '../components/main-layout/citation-footer/citation-footer.component';
 
 @Component({
   selector: 'app-main-layout',
